Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
psycopg2-binary
pandas
pdfminer.six
pytesseract
openai
python-dotenv
